refactor(media): share video description requests
